+++ cli/src/main.rs
@@ -1,7 +1,8 @@
 use std::path::PathBuf;
 
-use cargo_tangle::{create, deploy};
+use cargo_tangle::{create, deploy, keys};
 use clap::{Parser, Subcommand};
+use keys::KeyType;
 
 /// Tangle CLI tool
 #[derive(Parser, Debug)]
@@ -22,16 +23,18 @@ struct Cli {
 
 #[derive(Subcommand, Debug)]
 enum Commands {
-    /// Gadget subcommand
-    Gadget {
+    /// Blueprint subcommand
+    #[command(visible_alias = "bp")]
+    Blueprint {
         #[command(subcommand)]
         subcommand: GadgetCommands,
     },
 }
 
 #[derive(Subcommand, Debug)]
-enum GadgetCommands {
+pub enum GadgetCommands {
     /// Create a new blueprint
+    #[command(visible_alias = "c")]
     Create {
         /// The name of the blueprint
         #[arg(short, long)]
@@ -42,6 +45,7 @@ enum GadgetCommands {
     },
 
     /// Deploy a blueprint to the Tangle Network.
+    #[command(visible_alias = "d")]
     Deploy {
         /// HTTP RPC URL to use
         #[arg(long, value_name = "URL", default_value = "https://rpc.tangle.tools")]
@@ -52,6 +56,24 @@ enum GadgetCommands {
         /// The package to deploy (if the workspace has multiple packages).
         #[arg(short, long, value_name = "PACKAGE")]
         package: Option<String>,
+    },
+    /// Generate a key
+    Keygen {
+        /// The type of key to generate
+        #[arg(short, long, value_enum)]
+        key_type: KeyType,
+
+        /// The path to save the key (optional)
+        #[arg(short, long)]
+        path: Option<PathBuf>,
+
+        /// The SURI or seed to use for the generation of the key (optional)
+        #[arg(short, long)]
+        seed: Option<String>,
+
+        /// If true, the secret key will be printed along with the public key
+        #[arg(long)]
+        show_secret: bool,
     },
 }
 
@@ -76,7 +98,7 @@ async fn main() -> color_eyre::Result<()> {
     let cli = Cli::parse_from(args);
 
     match cli.command {
-        Commands::Gadget { subcommand } => match subcommand {
+        Commands::Blueprint { subcommand } => match subcommand {
             GadgetCommands::Create { name, source } => {
                 create::new_blueprint(name, source);
             }
@@ -98,6 +120,19 @@ async fn main() -> color_eyre::Result<()> {
                     signer_evm: None,
                 })
                 .await?;
+            }
+            GadgetCommands::Keygen {
+                key_type,
+                path,
+                seed,
+                show_secret,
+            } => {
+                keys::generate_key(
+                    key_type,
+                    path,
+                    seed.as_deref().map(str::as_bytes),
+                    show_secret,
+                )?;
             }
         },
     }