+++ macros/blueprint-proc-macro/src/event_listener/evm.rs
@@ -28,10 +28,27 @@ pub(crate) fn generate_evm_event_handler(
 ) -> TokenStream {
     let (instance_base, instance_name, instance_wrapper_name, _instance) =
         get_evm_instance_data(event_handler);
-    let event = event_handler.event().unwrap();
-    let event_converter = event_handler.event_converter().unwrap();
-    let callback = event_handler.callback().unwrap();
-    let abi_string = event_handler.abi().unwrap();
+    let event = event_handler
+        .get_event_listener()
+        .event
+        .as_ref()
+        .expect("Event type must be specified");
+    let event_converter = event_handler
+        .get_event_listener()
+        .pre_processor
+        .as_ref()
+        .unwrap();
+    let callback = event_handler
+        .get_event_listener()
+        .post_processor
+        .as_ref()
+        .unwrap();
+    let abi_string = event_handler
+        .get_event_listener()
+        .evm_args
+        .as_ref()
+        .and_then(|r| r.abi.clone())
+        .expect("ABI String must exist");
 
     quote! {
         #[derive(Debug, Clone)]