+++ sdk/src/lib.rs
@@ -65,6 +65,7 @@ pub mod job_runner;
 /// Gadget Context and context extensions
 pub mod ctx;
 
+pub mod docker;
 pub mod utils;
 
 // Re-exports
@@ -78,6 +79,7 @@ pub use structopt;
 pub use subxt_core;
 pub use tangle_subxt;
 pub use tokio;
+pub use uuid;
 
 // External modules usually used in proc-macro codegen.
 #[doc(hidden)]