+++ blueprint-test-utils/CHANGELOG.md
@@ -7,6 +7,89 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1](https://github.com/tangle-network/gadget/releases/tag/blueprint-test-utils-v0.1.1) - 2024-10-23
+
+### Added
+
+- add optional data dir to blueprint manager ([#342](https://github.com/tangle-network/gadget/pull/342))
+- eigenlayer incredible squaring blueprint and test ([#312](https://github.com/tangle-network/gadget/pull/312))
+
+### Fixed
+
+- *(sdk)* updated keystore support and fixes ([#368](https://github.com/tangle-network/gadget/pull/368))
+- *(sdk)* [**breaking**] downgrade substrate dependencies for now
+- add `data_dir` back to `GadgetConfiguration` ([#350](https://github.com/tangle-network/gadget/pull/350))
+
+### Other
+
+- release ([#378](https://github.com/tangle-network/gadget/pull/378))
+- release ([#362](https://github.com/tangle-network/gadget/pull/362))
+- Fix eigenlayer example ([#375](https://github.com/tangle-network/gadget/pull/375))
+- update to latest changes in tangle ([#367](https://github.com/tangle-network/gadget/pull/367))
+- Multi job runner + SDK main macro ([#346](https://github.com/tangle-network/gadget/pull/346))
+- Event Listener Upgrade + Wrapper Types + sdk::main macro ([#333](https://github.com/tangle-network/gadget/pull/333))
+- update naming ([#343](https://github.com/tangle-network/gadget/pull/343))
+- release ([#314](https://github.com/tangle-network/gadget/pull/314))
+- release ([#313](https://github.com/tangle-network/gadget/pull/313))
+- release ([#308](https://github.com/tangle-network/gadget/pull/308))
+- Remove Logger ([#311](https://github.com/tangle-network/gadget/pull/311))
+- Streamline keystore, cleanup testing, refactor blueprint manager, add tests, remove unnecessary code ([#285](https://github.com/tangle-network/gadget/pull/285))
+- CI Improvements ([#301](https://github.com/tangle-network/gadget/pull/301))
+- Expose executor from SDK ([#300](https://github.com/tangle-network/gadget/pull/300))
+- [MEGA PR] Overhaul repo, add Eigenlayer AVS example, remove many crates, add testing, remove unused code ([#246](https://github.com/tangle-network/gadget/pull/246))
+- Remove unused workspace dependencies ([#276](https://github.com/tangle-network/gadget/pull/276))
+- Add mpc blueprint starting point, cleanup abstractions ([#252](https://github.com/tangle-network/gadget/pull/252))
+- Add more checks to CI ([#244](https://github.com/tangle-network/gadget/pull/244))
+- Fix `too_long_first_doc_paragraph` ([#243](https://github.com/tangle-network/gadget/pull/243))
+- Promote all dependencies to workspace ([#233](https://github.com/tangle-network/gadget/pull/233))
+- Make `{core, io, common}` no_std and WASM compatible ([#231](https://github.com/tangle-network/gadget/pull/231))
+- Remove shell sdk and put inside blueprint manager ([#229](https://github.com/tangle-network/gadget/pull/229))
+- Blueprint testing ([#206](https://github.com/tangle-network/gadget/pull/206))
+
+## [0.1.1](https://github.com/tangle-network/gadget/releases/tag/blueprint-test-utils-v0.1.1) - 2024-10-23
+
+### Added
+
+- add optional data dir to blueprint manager ([#342](https://github.com/tangle-network/gadget/pull/342))
+- eigenlayer incredible squaring blueprint and test ([#312](https://github.com/tangle-network/gadget/pull/312))
+
+### Fixed
+
+- *(sdk)* updated keystore support and fixes ([#368](https://github.com/tangle-network/gadget/pull/368))
+- *(sdk)* [**breaking**] downgrade substrate dependencies for now
+- add `data_dir` back to `GadgetConfiguration` ([#350](https://github.com/tangle-network/gadget/pull/350))
+
+### Other
+
+- release ([#362](https://github.com/tangle-network/gadget/pull/362))
+- Fix eigenlayer example ([#375](https://github.com/tangle-network/gadget/pull/375))
+- update to latest changes in tangle ([#367](https://github.com/tangle-network/gadget/pull/367))
+- Multi job runner + SDK main macro ([#346](https://github.com/tangle-network/gadget/pull/346))
+- Event Listener Upgrade + Wrapper Types + sdk::main macro ([#333](https://github.com/tangle-network/gadget/pull/333))
+- update naming ([#343](https://github.com/tangle-network/gadget/pull/343))
+- release ([#314](https://github.com/tangle-network/gadget/pull/314))
+- release ([#313](https://github.com/tangle-network/gadget/pull/313))
+- release ([#308](https://github.com/tangle-network/gadget/pull/308))
+- Remove Logger ([#311](https://github.com/tangle-network/gadget/pull/311))
+- Streamline keystore, cleanup testing, refactor blueprint manager, add tests, remove unnecessary code ([#285](https://github.com/tangle-network/gadget/pull/285))
+- CI Improvements ([#301](https://github.com/tangle-network/gadget/pull/301))
+- Expose executor from SDK ([#300](https://github.com/tangle-network/gadget/pull/300))
+- [MEGA PR] Overhaul repo, add Eigenlayer AVS example, remove many crates, add testing, remove unused code ([#246](https://github.com/tangle-network/gadget/pull/246))
+- Remove unused workspace dependencies ([#276](https://github.com/tangle-network/gadget/pull/276))
+- Add mpc blueprint starting point, cleanup abstractions ([#252](https://github.com/tangle-network/gadget/pull/252))
+- Add more checks to CI ([#244](https://github.com/tangle-network/gadget/pull/244))
+- Fix `too_long_first_doc_paragraph` ([#243](https://github.com/tangle-network/gadget/pull/243))
+- Promote all dependencies to workspace ([#233](https://github.com/tangle-network/gadget/pull/233))
+- Make `{core, io, common}` no_std and WASM compatible ([#231](https://github.com/tangle-network/gadget/pull/231))
+- Remove shell sdk and put inside blueprint manager ([#229](https://github.com/tangle-network/gadget/pull/229))
+- Blueprint testing ([#206](https://github.com/tangle-network/gadget/pull/206))
+
+## [0.1.1](https://github.com/tangle-network/gadget/releases/tag/blueprint-test-utils-v0.1.1) - 2024-10-23
+
+### Fixed
+
+- *(sdk)* updated keystore support and fixes ([#368](https://github.com/tangle-network/gadget/pull/368))
+
 ## [0.1.1](https://github.com/tangle-network/gadget/releases/tag/blueprint-test-utils-v0.1.1) - 2024-09-30
 
 ### Other